Use integer math instead of float math
This patch replaces floating point arithmetic with integer arithmetic
to improve precision.  It exploits the arbitrary precision integers
that are now available in CircuitPython.  It removes the use of the
math.floor function which has known issues as described at:

https://github.com/adafruit/circuitpython/issues/572

The SI5351._PLL.frequency and SI5351._Clock.frequency properties are
now integers rather than floats. Strictly speaking this is a change
to the existing API.