Allow configuring with frequency values
Configuring divider values directly follows the datasheet, but it's
more convenient to think in terms of specific frequencies, especially
for output clocks.

For example, suppose you want a 33333.3333Hz (100000Hz/3) output.
Before it was necessary to do a bunch of calculations and then
write

    si.pll_a.configure_integer(20)
    si.clock_1.configure_fractional(si.pll_a, 468, 3, 4)
    si.clock_1.r_divider = R_DIV_32

With this change you can now write

    si.pll_a.configure_frequency(500000000)
    si.clock_1.configure_frequency(si.pll_a, 100000, 3)

For 32768Hz,

    si.clock_1.configure_frequency(si.pll_a, 32768)

Because the configuration process up to 3 arguments, I did not
make it a setter of the frequency property.